Comment Color Brightness
-> Please see arcticicestudio/nord#94 for all
   details about this design change decision.

Increased the comment color (`nord3`) brightness by 10% from a lightness
level of ~35% to ~45%.

This change also deprecates the "Custom Comment Contrast" theme setting.
It is not necessary anymore for users to increase the brightness on
their own when the default color has been increased by default.
A deprecation warning will be shown to notify all users who have set a
custom value for the configuration variable.

GH-60